# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOCUMENTATION DE L'API DU PROJET MESHUTE",
        "description": "MESHUTE est un projet de gestion de ressources immobilières qui a pour principal but de rendre leur accès facile",
        "contact": {
            "email": "knk.towork@gmail.com"
        },
        "license": {
            "name": "Apache2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Url de la documentation swagger Meshute"
        }
    ],
    "paths": {
...
        "/update-info/account/{id_compte}": {
            "put": {
                "tags": [
                    "Compte"
                ],
                "summary": "Mise à jour des informations personnelles de l'utilisateur",
                "description": "La gestion du compte utilisateur est principalement centrée sur la mise à jour des informations personnelles",
                "operationId": "updateAccountData",
                "parameters": [
                    {
                        "name": "id_compte",
                        "in": "path",
                        "description": "Identifiant du compte utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "date_naissance",
                                    "sexe",
                                    "telephone"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "date_naissance": {
                                        "type": "string"
                                    },
                                    "sexe": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "num_cni": {
                                        "type": "string"
                                    },
                                    "cni_recto": {
                                        "type": "string"
                                    },
                                    "cni_verso": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe renouvellé"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable"
                    }
                },
                "security": {
                    "sanctum": {
                        "0": []
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}